% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqMC.R
\name{update.seqMC}
\alias{update.seqMC}
\title{Update seqMC object after an observation (if y is not missing) or simply update the object to next time point
(if y is missing).}
\usage{
\method{update}{seqMC}(obj, y)
}
\arguments{
\item{obj}{seqMC object.}

\item{y}{vector or matrix, if y is a vector, it represents an observation at a single time step,
if y is a matrix, then each col is an observation at a time point, number of cols in y equal to number of 
time steps.}
}
\value{
seqMC object updated after given observation(s).
}
\description{
Update seqMC object after an observation (if y is not missing) or simply update the object to next time point
(if y is missing).
}
