% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqMC.R
\name{update.seqMC}
\alias{update.seqMC}
\title{Update seqMC object after an observation (if y is not missing) or simply update the object to next time point
(if y is missing).}
\usage{
\method{update}{seqMC}(obj, y)
}
\arguments{
\item{obj}{seqMC object.}

\item{y}{observation at time obj$t + 1.}
}
\value{
seqMC object updated to next time point.
}
\description{
Update seqMC object after an observation (if y is not missing) or simply update the object to next time point
(if y is missing).
}
